split static_condensation_utility.h into header and cpp
